cql3: in multi_item_terminal, return the vector of items by value
Returning by reference requires that the elements are internally stored in in
the multi_item_terminal as a std::vector, but in the next patch we will change
the internal type of lists::value from std::vector to utils::chunked_vector.

The copy is not a problem because all users of multi_item_terminal were copying
the returned vector.