compaction_manager: only start concurrent compactions when fan-in increases
Under default configuration, as soon as we have two similar-sized
sstables we can start compacting them. This is great for reducing
read amplification in a system under light write load, but it also
increases write amplification under heavy write load.

To reduce this write amplification, only admit new compactions if
they have higher fan-in than all current compactions. If the system
is incurring backlog, then the fan-in will increase until compactions
with retire with fan in at exactly the rate that compactions with the
same fan-in arrive. When write load reduces, the fan-in threshold
will reduce too.

This effectively trades write amplification against efficiency,
trying to self-tune by looking at the current fan-in as a measure
of backlog.