compaction_manager: Rework exception handling for regular compaction
Extracted most of the code from compaction_manager::maybe_stop_on_error(),
which will be eventually killed.

maybe_stop_on_error() is an attempt to generalize exception handling
in compaction manager but that kills readability and it's also ugly as
future has to be forwarded to it.

There's no semantic change in this patch. The code for conditionally
retrying compaction, when user stops it, was removed, but that
condition was always false for regular compaction.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>