hints: add sync_point_service
Adds a service which manages hint sync points.

A hint sync point is an object which allows tracking progress of hint
replay on multiple nodes at once. The service uses the same algorithm as
previosuly used for waiting for hint replay before repair.

The service provides methods for creating, waiting for and deleting hint
sync points. The service handles both the coordinator (node which tracks
progress on multiple nodes) and follower (node whose progress is
tracked) sides of the RPC.