conf/scylla.yaml: change default Prometheus listen address
Developers often run Scylla with the default conf/scylla.yaml provided
with the source distribution. The existing default listens for all ports
but one (19042, 10000, 9042, 7000) on the *localhost* IP address (127.0.0.1).
But just one port - 9180 (Prometheus metrics) - is listened on 0.0.0.0.
This patch changes the default to be 127.0.0.1 for port 9180 as well.

Note that this just changes the default scylla.yaml - users can still
choose whatever listening address they want by changing scylla.yaml
and/or passing command line parameters.

The benefits of this patch are:
1. More consistent.
2. Better security for developers (don't open ports on external
   addresses while testing).
3. Allow test/cql-pytest/run to run in parallel with a default run of
   Scylla (currently, it fails to run Scylla on a random IP address,
   because the default run of Scylla already took port 9180 on all IP
   addresses.

The third benefit is what led me to write this patch. Fixes #8757.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>
Message-Id: <20210530130307.906051-1-nyh@scylladb.com>