raft: fsm_test: test_leader_transferee_dies_upon_receiving_timeout_now
4-node cluster (A, B, C, D). A is initially elected a leader.
The leader adds a new configuration entry, that removes it from the
cluster (B, C, D).
Communicate the cluster up to the point where A starts to resign
its leadership (calls `transfer_leadership()`).
At this point, A should send a `timeout_now` message to one
the remaining nodes (B, C or D) and the new configuration should be
committed. But no nodes actually have received the `timeout_now` message
yet.

Determine on which node the message should arrive, accept the
`timeout_now` message and disconnect the target from the rest of the
group.

Check that after that the cluster, which has only two live members,
could progress and elect a new leader through a normal election process.

tests: unit(dev, debug)

Signed-off-by: Pavel Solodovnikov <pa.solodovnikov@scylladb.com>